Add frontend refactor
