+++ client/src/CodeClaimPage/templateClaim/TemplateClaimLoading.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 // assets
-import { ReactComponent as LoadingIcon } from '../../images/loading.svg';
+import { ReactComponent as LoadingIcon } from 'images/loading.svg';
 
 export const TemplateClaimLoading = () => (
   <div className="template_claim_loading">
