+++ client/src/api.ts
@@ -1,7 +1,6 @@
 import queryString from 'query-string';
 
 import { authClient } from './auth';
-import { Template } from './templates/TemplatePage/types';
 
 export type Address = string;
 
@@ -9,8 +8,7 @@ export type Params = {
   [key: string]: string | number | boolean | undefined;
 };
 export interface TemplatesResponse<Result> {
-  count: number;
-
+  total: number;
   next?: string;
   previous?: string;
   event_templates: Result[];
@@ -57,6 +55,45 @@ export interface ClaimProof {
   proof: string;
 }
 
+export type Template = {
+  id: number;
+  name: string;
+  title_image: string;
+  title_link: string;
+  header_link_text: string;
+  header_link_url: string;
+  header_color: string;
+  header_link_color: string;
+  main_color: string;
+  footer_color: string;
+  left_image_url: string;
+  left_image_link: string;
+  right_image_url: string;
+  right_image_link: string;
+  mobile_image_url: string;
+  mobile_image_link: string;
+  footer_icon: string;
+  secret_code: string;
+};
+export type TemplatePageFormValues = {
+  name: string;
+  title_image: Blob | string;
+  title_link: string;
+  header_link_text: string;
+  header_link_url: string;
+  header_color: string;
+  header_link_color: string;
+  main_color: string;
+  footer_color: string;
+  left_image_url: Blob | string;
+  left_image_link: string;
+  right_image_url: Blob | string;
+  right_image_link: string;
+  mobile_image_url: Blob | string;
+  mobile_image_link: string;
+  footer_icon: Blob | string;
+  secret_code: string;
+};
 export type EventTemplate = {
   created_date: string;
   footer_color: string;
@@ -259,14 +296,17 @@ export async function getEvents(): Promise<PoapEvent[]> {
 
 export type TemplateResponse = TemplatesResponse<Template>;
 
-export async function getTemplates({ limit = 10, page = 1, name = '' }: Params = {}): Promise<
+export async function getTemplates({ limit = 10, offset = 0, name = '' }: Params = {}): Promise<
   TemplateResponse
 > {
-  return fetchJson(`${API_BASE}/event-templates?limit=${limit}&offset=${page}&name=${name}`);
+  return fetchJson(`${API_BASE}/event-templates?limit=${limit}&offset=${offset}&name=${name}`);
 }
 
 export async function getTemplateById(id?: number): Promise<Template> {
-  return fetchJson(`${API_BASE}/event-templates/${id}`);
+  const isAdmin = authClient.isAuthenticated();
+  return isAdmin
+    ? secureFetch(`${API_BASE}/event-templates-admin/${id}`)
+    : fetchJson(`${API_BASE}/event-templates/${id}`);
 }
 
 export async function getEvent(fancyId: string): Promise<null | PoapFullEvent> {
@@ -409,14 +449,14 @@ export async function createEvent(event: FormData) {
   });
 }
 
-export async function createTemplate(event: FormData) {
+export async function createTemplate(event: FormData): Promise<Template> {
   return fetchJson(`${API_BASE}/event-templates`, {
     method: 'POST',
     body: event,
   });
 }
 
-export async function updateTemplate(event: FormData, id: number) {
+export async function updateTemplate(event: FormData, id: number): Promise<void>  {
   return fetchJsonNoResponse(`${API_BASE}/event-templates/${id}`, {
     method: 'PUT',
     body: event,
