+++ client/src/backoffice/Main.tsx
@@ -4,12 +4,12 @@ import { Link, Redirect, Route, withRouter, Switch } from 'react-router-dom';
 import { slide as Menu } from 'react-burger-menu';
 
 // lib
-import { AuthContext, authClient } from '../auth';
+import { AuthContext, authClient } from 'auth';
 
 /* Assets */
-import PoapLogo from '../images/POAP.svg';
-import Calendar from '../images/calendar.svg';
-import Qr from '../images/qr-code.svg';
+import PoapLogo from 'images/POAP.svg';
+import Calendar from 'images/calendar.svg';
+import Qr from 'images/qr-code.svg';
 
 /* Constants */
 import { ROUTES, LABELS } from '../lib/constants';
@@ -23,7 +23,6 @@ import { InboxPage } from './InboxPage';
 import { InboxListPage } from './InboxListPage';
 import { QrPage } from './QrPage';
 import { EventsPage } from './EventsPage';
-// import { EventList, CreateEventForm, EditEventForm } from './EventsPage';
 
 export const MintersPage = () => <div> This is a MintersPage </div>;
 
@@ -107,7 +106,7 @@ export const NavigationMenu = withRouter(({ history }) => {
       <SidebarLink route={ROUTES.events} handleClick={closeMenu} />
 
       <SidebarLink route={ROUTES.qr} handleClick={closeMenu} />
-      
+
       <SidebarLink route={ROUTES.template} handleClick={closeMenu} />
 
       {!isAdmin && <SidebarLink route={ROUTES.adminLogin} handleClick={closeMenu} />}
@@ -135,12 +134,12 @@ const Landing = () => {
         <h3>Manage Events</h3>
         <img className={'icon'} src={Calendar} alt={'Manage Events'} />
       </Link>
-      {isAdmin &&
+      {isAdmin && (
         <Link to={ROUTES.qr.path} className={'card card-link'}>
           <h3>Manage QR Codes</h3>
           <img className={'icon'} src={Qr} alt={'Manage QR Codes'} />
         </Link>
-      }
+      )}
     </div>
   );
 }
